test: 데이터베이스 연결 테스트 디렉토리 이동
