+++ src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -1,11 +1,5 @@
 package roomescape;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -16,17 +10,5 @@ import org.springframework.test.annotation.DirtiesContext;
 public class RoomescapeApplicationTest {
 
     @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    @Test
-    @DisplayName("데이터베이스 연결은 성공한다.")
-    public void 데이터베이스_연결은_성공한다() {
-        try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
-            assertThat(connection).isNotNull();
-            assertThat(connection.getCatalog()).isEqualTo("DATABASE");
-            //assertThat(connection.getMetaData().getTables(null, null, "RESERVATION", null).next()).isTrue(); //FIXME
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
+    protected JdbcTemplate jdbcTemplate;
 }
