+++ src/test/java/roomescape/domain/reservation/DatabaseTest.java
@@ -1,0 +1,24 @@
+package roomescape.domain.reservation;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import roomescape.RoomescapeApplicationTest;
+
+public class DatabaseTest extends RoomescapeApplicationTest {
+
+    @Test
+    @DisplayName("데이터베이스 연결은 성공한다.")
+    public void 데이터베이스_연결은_성공한다() {
+        try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
+            assertThat(connection).isNotNull();
+            assertThat(connection.getCatalog()).isEqualTo("DATABASE");
+            assertThat(connection.getMetaData().getTables(null, null, "RESERVATION", null).next()).isTrue();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
